+++ backend/src/main/java/com/showing/backend/api/service/RankingService.java
@@ -2,6 +2,6 @@ package com.showing.backend.api.service;
 
 public interface RankingService {
 
-    float getStarPointAverage(Long performanceId);
+    double getStarPointAverage(Long performanceId);
 
 }
